+++ chaosmagpy/__init__.py
@@ -8,4 +8,4 @@ __all__ = ["CHAOS", "load_CHAOS_matfile", "load_CHAOS_shcfile",
            "load_CovObs_txtfile", "load_gufm1_txtfile",
            "basicConfig", "synth_values"]
 
-__version__ = "0.4"
+__version__ = "0.5dev"